Move implementation of Codec.Archive.Tar.Index to an internal module to be able to export more stuff for tests
